refactor: rename package `events` to `event` and `sinks` to `sink`
