+++ dwcj-addons/dwcj-googlecharts/src/main/java/org/dwcj/addons/googlecharts/events/GoogleChartReadyEvent.java
@@ -2,7 +2,7 @@ package org.dwcj.addons.googlecharts.events;
 
 import java.util.Map;
 import org.dwcj.addons.googlecharts.GoogleChart;
-import org.dwcj.component.events.Event;
+import org.dwcj.component.event.Event;
 import org.dwcj.component.webcomponent.annotations.EventExpressions;
 import org.dwcj.component.webcomponent.annotations.EventName;
