+++ dwcj-addons/dwcj-googlecharts/src/main/java/org/dwcj/addons/googlecharts/events/GoogleChartSelectedEvent.java
@@ -4,7 +4,7 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import java.util.Map;
 import org.dwcj.addons.googlecharts.GoogleChart;
-import org.dwcj.component.events.Event;
+import org.dwcj.component.event.Event;
 import org.dwcj.component.webcomponent.annotations.EventExpressions;
 import org.dwcj.component.webcomponent.annotations.EventName;
