+++ dwcj-engine/src/main/java/org/dwcj/component/applayout/AppLayout.java
@@ -6,7 +6,7 @@ import org.dwcj.component.HasStyle;
 import org.dwcj.component.HasVisibility;
 import org.dwcj.component.applayout.event.AppLayoutDrawerCloseEvent;
 import org.dwcj.component.applayout.event.AppLayoutDrawerOpenEvent;
-import org.dwcj.component.events.EventListener;
+import org.dwcj.component.event.EventListener;
 import org.dwcj.component.webcomponent.PropertyDescriptor;
 import org.dwcj.component.webcomponent.WebComponent;
 import org.dwcj.component.webcomponent.annotations.NodeAttribute;
