+++ dwcj-engine/src/main/java/org/dwcj/component/button/event/ButtonClickEvent.java
@@ -2,7 +2,7 @@ package org.dwcj.component.button.event;
 
 import java.util.Map;
 import org.dwcj.component.button.Button;
-import org.dwcj.component.events.Event;
+import org.dwcj.component.event.Event;
 
 /**
  * A push, or click, event for the Button class.
