+++ dwcj-engine/src/main/java/org/dwcj/component/webcomponent/WebComponent.java
@@ -24,9 +24,9 @@ import org.dwcj.annotation.InlineStyleSheet;
 import org.dwcj.annotation.JavaScript;
 import org.dwcj.component.AbstractComponent;
 import org.dwcj.component.AbstractDwcComponent;
-import org.dwcj.component.events.Event;
-import org.dwcj.component.events.EventDispatcher;
-import org.dwcj.component.events.EventListener;
+import org.dwcj.component.event.Event;
+import org.dwcj.component.event.EventDispatcher;
+import org.dwcj.component.event.EventListener;
 import org.dwcj.component.htmlcontainer.HtmlContainer;
 import org.dwcj.component.htmlcontainer.event.HtmlContainerJavascriptEvent;
 import org.dwcj.component.webcomponent.annotations.EventExpressions;
